+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Note
 
-## v0.3.0
+## v0.2.2
 
 - Supported `preferences API methods`. ([#19](https://github.com/mastodon-dart/mastodon-api/issues/19))
   - `GET /api/v1/preferences`